+++ test/support/piano.js
@@ -1,0 +1,10 @@
+
+if (typeof(MIDI) === 'undefined') var MIDI = {};
+if (typeof(MIDI.Soundfont) === 'undefined') MIDI.Soundfont = {};
+MIDI.Soundfont.acoustic_grand_piano = {
+  "A0": "data:audio/ogg;base64,MQ==",
+  "Bb0": "data:audio/ogg;base64,MQ==",
+  "B0": "data:audio/ogg;base64,MQ==",
+  "C1": "data:audio/ogg;base64,MQ==",
+  "Db1": "data:audio/ogg;base64,MQ==",
+}